Another round of fixes and standardizations

+++ 1-Elementary-Algorithms/1.3.5_Minimizing_Late_Jobs_(Moore-Hodgson).cpp
@@ -39,7 +39,7 @@ int maximize_on_time_jobs(std::vector<TimedJob> jobs) {
       durations.pop();
     }
   }
-  return durations.size();
+  return static_cast<int>(durations.size());
 }
 
 /*** Example Usage ***/
